IMA: use vfs_getattr_nosec to get the i_version
[ Upstream commit db1d1e8b9867aae5c3e61ad7859abfcc4a6fd6c7 ]

IMA currently accesses the i_version out of the inode directly when it
does a measurement. This is fine for most simple filesystems, but can be
problematic with more complex setups (e.g. overlayfs).

Make IMA instead call vfs_getattr_nosec to get this info. This allows
the filesystem to determine whether and how to report the i_version, and
should allow IMA to work properly with a broader class of filesystems in
the future.

Reported-and-Tested-by: Stefan Berger <stefanb@linux.ibm.com>
Reviewed-by: Christian Brauner <brauner@kernel.org>
Signed-off-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>
Stable-dep-of: be84f32bb2c9 ("ima: Fix use-after-free on a dentry's dname.name")
Signed-off-by: Sasha Levin <sashal@kernel.org>